separated the checkoutpage in different components
